Module aeb_rs::grid

Generic deterministic occupancy grid and associated utilities

Structs

Grid

An nxn occupancy grid.

GridPoint

An index indo the grid, garrentied to be in bounds. We define a point on the grid to be the top left corner of the grid squares if a grid is drawn out. The top left of the grid is (0,0).

GridPointf

A point in gridspace, not aligned to the grid nor bounds checked.

GridSize

The n in nxn grid. This is garrentied to be odd and < f32::MAX.

KartPoint

A point in the reference frame of the kart, where (0,0) is the location of the front range sensor.

Enums

Cell

A cell in the grid

GridErr

Grid associated errors

Type Definitions

Line

A line given by two vertices. The order of vertices does not matter.