Struct aeb_rs::Aeb

pub struct Aeb<const GRID_N: usize> { /* private fields */ }

The entrypoint to the automatic emergency braking algorithm.

Implementations

impl<const GRID_N: usize> Aeb<GRID_N>

pub fn new(
start_vel: f32,
start_wheel_angle: f32,
wheelbase: f32,
collision_box: ((f32, f32), (f32, f32)),
axel_to_sensor: KartPoint,
min_ttc: f32
) -> Self

Configures AEB.

Config

start_vel: starting velocity, in m/s

start_wheel_angle: starting ackermann wheel angle, in degrees, positive is right.

wheelbase: distance between axles, in meters.

collision_box: collision bounding box, defined as the top left and bottom right points relative to the center of the rear axel. This is in normal coordinates.

axel_to_sensor: The transform of the axel to the front range sensor, in the kart frame.

min_ttc: the minimum allowed time to collision.

pub fn collision_check(&mut self, points: Option<&[KartPoint]>) -> (bool, usize)

Runs the collision checking algorithm with the current parameters. The passed points will also be added to the occupancy grid, if desired. After this call, the grid will be cleared.

Returns true if the vehicle should brake, along with it’s collision time in ms.

pub fn predict_pos(&self, t: usize) -> (KartPoint, f32)

Predicts the future position of the kart given the currently configured params, at some time.

Returns final (position, heading), where heading is in rad and negative if left.

Args

t: time in ms

pub fn create_obb(
&self,
pos: KartPoint,
yaw: f32
) -> [((f32, f32), (f32, f32)); 4]

Creates the oriented bounding box given the karts position and yaw in rad.

This box may lie outside the grid partially or totally.

pub fn update_velocity(&mut self, velocity: f32)

Updates the current velocity, in m/s.

pub fn update_steering(&mut self, steering_angle: f32)

Updates the current ackermann steering angle, in degrees.

pub fn update_ttc(&mut self, min_ttc: f32)

Updates the minimum allowed time to collision, in seconds.

The minimum time to collision will be the boundary that we will collision detect to, ie the minimum tolerated collision time.

pub fn add_points(&mut self, points: &[KartPoint])

Adds points to the grid to be used during the next collision check.

Points that are < 15cm from the data source will be filtered out.